Fix a couple issues with the rotation of the laser altimeter.  It now correctly honors the setting in the config file.

+++ Unreal/Plugins/AirSim/Source/VehiclePawnWrapper.h
@@ -90,7 +90,7 @@ public: //interface
     std::string getVehicleConfigName() const;
 
     int getRemoteControlID() const;
-    int getLidarForwardPitch() const;
+    float getLidarForwardPitch() const;
 
 protected:
     UPROPERTY(VisibleAnywhere)
